+++ R/games_info.r
@@ -13,6 +13,7 @@
 #' nrow(domino_games())
 #' nrow(go_games())
 #' nrow(marble_games())
+#' nrow(morris_games())
 #' nrow(piecepack_games())
 #' nrow(stackpack_games())
 #' @return A [tibble::tibble()] data frame with character columns "game", "methods", "comment", and "url"
@@ -61,6 +62,13 @@ go_games <- function() {
 #' @export
 marble_games <- function() {
     bind_rows(marble_games_variant()) %>%
+        arrange_games()
+}
+
+#' @rdname games_info
+#' @export
+morris_games <- function() {
+    bind_rows(morris_games_variant()) %>%
         arrange_games()
 }